+++ contracts/Interfaces.sol
@@ -23,7 +23,6 @@ import "@openzeppelin/contracts/access/Ownable.sol";
 import "@chainlink/contracts/src/v0.6/dev/AggregatorInterface.sol";
 import "@uniswap/v2-periphery/contracts/interfaces/IUniswapV2Router01.sol";
 
-
 interface ILiquidityPool {
     event Withdraw(
         address indexed account,
@@ -32,25 +31,29 @@ interface ILiquidityPool {
     );
 
     event Provide(address indexed account, uint256 amount, uint256 writeAmount);
+
     function lock(uint256 amount) external;
+
     function unlock(uint256 amount) external;
+
     function unlockPremium(uint256 amount) external;
+
     function send(address payable account, uint256 amount) external;
-    function setLockupPeriod(uint value) external;
+
+    function setLockupPeriod(uint256 value) external;
+
     function totalBalance() external view returns (uint256 amount);
 }
 
-
 interface IERCLiquidityPool is ILiquidityPool {
     function sendPremium(uint256 amount) external;
+
     function token() external view returns (IERC20);
 }
-
 
 interface IETHLiquidityPool is ILiquidityPool {
     function sendPremium() external payable;
 }
-
 
 // for the future
 // interface ERC20Incorrect {
